# server-management-system/api-gateway/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Server Management System API
  description: |
    API quản lý 10.000 server.
    
    ## Tính năng chính
    - **Auth**: Đăng ký, đăng nhập, JWT, RBAC (Admin / Operator / Viewer)
    - **Server CRUD**: Tạo, xem (filter/sort/paging), cập nhật, xóa
    - **Import/Export Excel**: Upload .xlsx để import hàng loạt, export ra file
    - **Monitoring**: Health-check TCP 10.000 server mỗi 60s, ghi Elasticsearch
    - **Báo cáo**: Báo cáo định kỳ qua Email + API báo cáo chủ động
    
    ## Authentication
    Tất cả API (trừ register, login, refresh) yêu cầu JWT Bearer token.
    Mỗi API có scope riêng (xem từng endpoint).
    
    ## Error Format
    ```json
    {
      "status": "error",
      "code": 42201,
      "message": "Validation failed",
      "errors": [{ "field": "ipv4", "code": "INVALID_FORMAT", "message": "..." }],
      "meta": { "request_id": "req-uuid", "timestamp": "2026-06-10T10:00:00Z" }
    }
    ```
  version: 1.0.0
  contact:
    name: SMS Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token. Lấy từ POST /auth/login.
        Token hết hạn sau 15 phút, dùng POST /auth/refresh để làm mới.

  # ═══════════════════════════════════════
  # Schemas
  # ═══════════════════════════════════════
  schemas:
...
    ImportJobResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        status:
          type: string
          enum: [pending, processing, completed, failed]
          example: pending
        file_name:
          type: string
          example: servers.xlsx
        message:
          type: string
          example: "File đã được nhận, đang xử lý bất đồng bộ"
    ImportJobStatusResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
          example: completed
        file_name:
          type: string
        total_rows:
          type: integer
          example: 500
        success_count:
          type: integer
          example: 480
        failed_count:
          type: integer
          example: 20
        success_list:
          type: array
          items:
            type: object
            properties:
              server_id:
                type: string
              server_name:
                type: string
        failed_list:
          type: array
          items:
            type: object
            properties:
              row_number:
                type: integer
              server_id:
                type: string
              server_name:
                type: string
              error_reason:
                type: string
        error_message:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
        meta:
          $ref: '#/components/schemas/MetaInfo'
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
          example: 42201
        message:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        meta:
          $ref: '#/components/schemas/MetaInfo'
    FieldError:
      type: object
      properties:
        field:
          type: string
          example: ipv4
        code:
          type: string
          example: INVALID_FORMAT
        message:
          type: string
          example: "IPv4 must be in format x.x.x.x"
    MetaInfo:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440099"
        timestamp:
          type: string
          format: date-time
          example: "2026-06-10T10:00:00Z"

# ═══════════════════════════════════════
# Paths
# ═══════════════════════════════════════
paths:
...
  /servers/import:
    post:
      tags: [Import/Export]
      summary: Import danh sách server từ file Excel
      description: |
        Upload file .xlsx. Xử lý bất đồng bộ qua Kafka.
        Bỏ qua các `server_id` hoặc `server_name` đã tồn tại.
        **Scope:** `server:import`
        **Role:** Admin only
      operationId: importServers
      security:
        - BearerAuth: [server:import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: File Excel .xlsx
      responses:
        '202':
          description: File đã nhận, đang xử lý
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ImportJobResponse'
        '400':
          description: File không hợp lệ (sai format, không phải .xlsx)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Thiếu JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Không có scope `server:import`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /servers/import/{job_id}:
    get:
      tags: [Import/Export]
      summary: Kiểm tra tiến độ job import
      description: |
        **Scope:** `server:import`
        **Role:** Admin only
      operationId: getImportStatus
      security:
        - BearerAuth: [server:import]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        '200':
          description: Trạng thái job import
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ImportJobStatusResponse'
        '401':
          description: Thiếu JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Không có scope `server:import`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job không tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...